@@ -32,6 +32,7 @@
     "@mui/lab": "^5.0.0-alpha.48",
     "@mui/material": "^5.0.1",
     "@mui/styles": "^5.0.1",
+    "@mui/x-date-pickers": "^5.0.0-beta.2",
     "apollo-cache-inmemory": "^1.6.5",
     "apollo-client": "^2.6.8",
     "apollo-link": "^1.2.13",
@@ -66,10 +67,10 @@
     "qs": "^6.10.1",
     "random-color-rgb": "^1.1.1",
     "rc-slider": "^9.7.2",
-    "react": "~17.0.2",
+    "react": "^18.2.0",
     "react-apollo": "^3.1.3",
     "react-autosuggest": "^10.1.0",
-    "react-dom": "~17.0.2",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.9",
     "react-ga": "^3.3.0",
     "react-helmet": "^6.1.0",
@@ -122,10 +123,10 @@
     "standard": "^16.0.3"
   },
   "resolutions": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "**/react": "17.0.2",
-    "**/react-dom": "17.0.2"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "**/react": "18.2.0",
+    "**/react-dom": "18.2.0"
   },
   "lint-staged": {
     "*.js": [
