@@ -1,7 +1,6 @@
 import React from 'react'
 import { makeStyles } from '@mui/styles'
 import CircularProgress from '@mui/material/CircularProgress'
-import Box from '@mui/material/Box'
 
 import styles from './styles'
 
@@ -11,9 +10,9 @@ const Loader = () => {
   const classes = useStyles()
 
   return (
-    <Box className={classes.root}>
+    <div className={classes.root}>
       <CircularProgress m={2} color='secondary' />
-    </Box>
+    </div>
   )
 }
 
