@@ -2,7 +2,6 @@ import React, { memo } from 'react'
 import Typography from '@mui/material/Typography'
 import { makeStyles } from '@mui/styles'
 import { useTranslation } from 'react-i18next'
-import Box from '@mui/material/Box'
 
 import TitlePage from '../../components/PageTitle'
 
@@ -15,11 +14,11 @@ const About = () => {
   const { t } = useTranslation('about')
 
   return (
-    <Box>
+    <div>
       <TitlePage title={t('tabTitle')} />
-      <Box>
-        <Box className={classes.firstContainer}>
-          <Box className={classes.wrapperContainers}>
+      <div>
+        <div className={classes.firstContainer}>
+          <div className={classes.wrapperContainers}>
             <Typography variant='h5'>{t('title')}</Typography>
             <Typography variant='h6'>{t('subtitle1')}</Typography>
             <Typography variant='body2' align='justify' paragraph>
@@ -31,11 +30,11 @@ const About = () => {
             <Typography variant='body2' align='justify' paragraph>
               {t('body1.paragraph3')}
             </Typography>
-          </Box>
-        </Box>
+          </div>
+        </div>
 
-        <Box className={classes.lastContainer}>
-          <Box className={classes.wrapperContainers}>
+        <div className={classes.lastContainer}>
+          <div className={classes.wrapperContainers}>
             <Typography variant='h6'>{t('subtitle2')}</Typography>
             <Typography variant='body2' align='justify' paragraph>
               {t('body2.paragraph1')}
@@ -46,11 +45,11 @@ const About = () => {
             <Typography variant='body2' align='justify' paragraph>
               {t('body2.paragraph3')}
             </Typography>
-          </Box>
-        </Box>
+          </div>
+        </div>
 
-        <Box className={classes.middleContainer}>
-          <Box className={classes.wrapperContainers}>
+        <div className={classes.middleContainer}>
+          <div className={classes.wrapperContainers}>
             <Typography variant='h6'>{t('subtitle3')}</Typography>
             <Typography variant='body2' align='justify' paragraph>
               {t('body3.paragraph1')}
@@ -103,10 +102,10 @@ const About = () => {
               src='/EOS-Rate-Infographic.jpg'
               alt='EOS block producer ratings'
             />
-          </Box>
-        </Box>
-      </Box>
-    </Box>
+          </div>
+        </div>
+      </div>
+    </div>
   )
 }
 
