@@ -3,7 +3,6 @@ import { makeStyles } from '@mui/styles'
 import { useTranslation } from 'react-i18next'
 import { Tweet } from 'react-twitter-widgets'
 import Typography from '@mui/material/Typography'
-import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import Link from '@mui/material/Link'
 import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos'
@@ -35,12 +34,12 @@ const Home = () => {
   }, [])
 
   return (
-    <Box>
+    <div>
       <TitlePage title={t('title')} />
-      <Box className={classes.mainCoverContainer}>
+      <div className={classes.mainCoverContainer}>
         <LogoSvg />
-        <Box className={classes.headerBg}>
-          <Box className={classes.gradientBg}>
+        <div className={classes.headerBg}>
+          <div className={classes.gradientBg}>
             <Typography className={classes.headerTitle}>
               {t('knowBP')}
             </Typography>
@@ -60,14 +59,14 @@ const Home = () => {
             >
               {t('startRating')}
             </Button>
-          </Box>
-        </Box>
-      </Box>
-      <Box className={classes.videoBox}>
+          </div>
+        </div>
+      </div>
+      <div className={classes.videoBox}>
         <Typography className={classes.videoTitle}>{t('tour')}</Typography>
-        <Box className={classes.video}>
+        <div className={classes.video}>
           <Video src={t('videoUrl')} />
-        </Box>
+        </div>
         <Link
           variant='body2'
           href='/about'
@@ -76,8 +75,8 @@ const Home = () => {
         >
           {t('learnMore')} <ArrowForwardIosIcon />
         </Link>
-      </Box>
-      <Box className={classes.blockProducers}>
+      </div>
+      <div className={classes.blockProducers}>
         <Typography className={classes.votes}>{t('mostVoted')}</Typography>
         <Link
           variant='body2'
@@ -86,11 +85,11 @@ const Home = () => {
         >
           {t('bpList')} <ArrowForwardIosIcon />
         </Link>
-      </Box>
-      <Box className={classes.wrapperGrid}>
-        <Box className={classes.gridRow}>
+      </div>
+      <div className={classes.wrapperGrid}>
+        <div className={classes.gridRow}>
           {(state.homeProducers.data || []).map((blockProducer, index) => (
-            <Box
+            <div
               className={classes.gridItem}
               key={`${blockProducer.owner}-main-block-card`}
               id={`cardBox${index}`}
@@ -104,9 +103,9 @@ const Home = () => {
                 pathLink='block-producers'
                 buttonLabel={t('addToVote')}
                 average={getAverageValue(
-                  _get(blockProducer, 'totalStats.average', 0)
+                  _get(blockProducer, 'total_average', 0)
                 )}
-                rate={_get(blockProducer, 'totalStats.ratings_cntr', 0)}
+                rate={_get(blockProducer, 'ratings_cntr', 0)}
                 isNewRate={
                   state.user &&
                   state.user.userData.userRates.some(
@@ -114,11 +113,11 @@ const Home = () => {
                   )
                 }
               />
-            </Box>
+            </div>
           ))}
-        </Box>
-      </Box>
-      <Box className={clsx(classes.blockProducers, classes.twitter)}>
+        </div>
+      </div>
+      <div className={clsx(classes.blockProducers, classes.twitter)}>
         <Typography className={classes.votes}>{t('mention')}</Typography>
         <Link
           variant='body2'
@@ -128,21 +127,21 @@ const Home = () => {
         >
           {t('moreNews')} <ArrowForwardIosIcon />
         </Link>
-      </Box>
-      <Box className={clsx(classes.wrapperGrid, classes.twitter)}>
-        <Box className={classes.gridRow}>
-          <Box className={classes.gridItem}>
+      </div>
+      <div className={clsx(classes.wrapperGrid, classes.twitter)}>
+        <div className={classes.gridRow}>
+          <div className={classes.gridItem}>
+            <Tweet tweetId='1563298151556976640' />
+          </div>
+          <div className={classes.gridItem} id='cardBox1'>
             <Tweet tweetId='1460649455443759115' />
-          </Box>
-          <Box className={classes.gridItem} id='cardBox1'>
+          </div>
+          <div className={classes.gridItem} id='cardBox2'>
             <Tweet tweetId='1451875419246968833' />
-          </Box>
-          <Box className={classes.gridItem} id='cardBox2'>
-            <Tweet tweetId='1457795829666697217' />
-          </Box>
-        </Box>
-      </Box>
-    </Box>
+          </div>
+        </div>
+      </div>
+    </div>
   )
 }
 
